Backend/requirements.txt
tensorflow
flask==2.3.3
flask-cors==4.0.0
pillow==10.0.0
numpy==1.24.3
opencv-python==4.8.1.78
python-dotenv==1.0.0
gunicorn==21.2.0